+++ docs/source/changes.rst
@@ -37,6 +37,7 @@ Release 0.5.0 - (under development)
 * Enable specification of memory pool allocator to linalg functions (enh.  by 
   Thomas Unterthiner).
 * Improve misc.select_block_grid_sizes() logic to handle different GPU hardware.
+* Compute transpose using CUDA 5.0 CUBLAS functions rather than with inefficient naive kernel.
 
 Release 0.042 - (March 10, 2013)
 --------------------------------